requirements.txt
requests~=2.31.0
python-dotenv~=1.1.1
python-telegram-bot~=22.2
pillow~=11.3.0